Add fist draft test
